My SEPTA Calculator Files
This commit adds the necessary files that support my SEPTA calculator.
The calculator's functionality is created using jquery, which is found
at the bottom of the index page.  It is styled using the core.css.  I
have also included a minified copy of jquery and the json data file.
Thank you in advance for reviewing my code.  Take care and have a good
day.