Scott Rockers Code Challenge
Hey TJ, here are my files for the code challenge. I worked off and on
for a couple of hours last night while getting the kids down. This
project was a fun little challenge, it had my mind going for a while
after finishing.  I wasn't able to Q&A accessibility within the allotted
time frame but all the functionality and responsive stuff is there and
working. You can view it in action here:

http://scottrockers.com/dev/thinkbrownstone/code-challenges/septa-fare-calculator/