Andrew Gladish Code Challenge
Made this in AngularJS 1.6.6 so it needs to be hosted on a server, local or otherwise. Then navigate to the address it is being hosted on.

Since this was meant to be a widget I build it as a directive that could easily be plugged into any angular app so long as the server call gets the right information. Wanted to build out some error catching functionality but time was running low.

Tested on Chrome, Firefox, Android Chrome, iPhone Safari, IE8, IE11, and Edge. Couldn't get access to a machine with MacOS Safari so I apologize if there are issues there.

I am excited to hear your thoughts and feedback on my submission. Please let me know if you have any questions or concerns with it.

Thanks,
Andrew